requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
attrs==17.4.0
bandit==1.4.0
cached-property==1.4.2
certifi==2018.1.18
celery>=4.2,<=4.3
cffi==1.11.5
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cryptography==2.1.4
dateutils==0.6.6
decorator==4.2.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.4
django-choices==1.6.0
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-extensions==2.0.6
django-filter==1.1.0
django-rest-swagger==2.1.2
django-simple-history==2.0
django-sslserver==0.21
djangorestframework==3.8.2
djangorestframework-csv==2.1.0
djangorestframework-expiring-authtoken==0.1.4
djangorestframework-filters==0.10.2
djangosaml2==0.16.11
drf-extensions==0.3.1
freezegun==0.3.10
future==0.16.0
gitdb2==2.0.3
GitPython==2.1.7
greenlet==0.4.13
idna==2.6
isodate==0.6.0
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.10
locustio==0.8.1
lxml==4.2.1
MarkupSafe==1.0
maya==0.6.1
model-mommy==1.5.1
more-itertools==4.1.0
msgpack-python==0.5.6
openapi-codec==1.3.2
Paste==2.0.3
pbr==4.0.1
pluggy==0.6.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
py==1.5.3
pycparser==2.18
pycryptodomex==3.6.0
pyflakes==1.6.0
PyJWT==1.7.1
pyOpenSSL==17.5.0
pysaml2==4.4.0
pytest==3.5.0
pytest-cov==2.5.1
pytest-django==3.1.2
python-dateutil==2.7.2
pytz==2018.3
pyzmq==17.0.0
repoze.who==2.3
requests==2.18.4
requests-toolbelt==0.8.0
simplejson==3.13.2
six==1.11.0
smmap2==2.0.3
sqlparse==0.2.4
stevedore==1.28.0
typing==3.6.4
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
WebOb==1.8.0
Werkzeug==0.14.1
zeep==2.5.0
zope.interface==4.4.3
